+++ src/main/kotlin/com/pafoid/skate/engine/ecs/components/NonPickable.kt
@@ -1,3 +1,6 @@
 package com.pafoid.skate.engine.ecs.components
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 class NonPickable: Component()